requirements2.txt
appdirs==1.4.4
astunparse==1.6.3
autograd==1.8.0
autoray==0.7.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
diastatic-malt==2.15.2
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gast==0.6.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pennylane==0.42.3
pennylane_lightning==0.42.0
pillow==11.3.0
pip==23.2.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
requests==2.32.5
rustworkx==0.17.1
scipy==1.16.2
scipy-openblas32==0.3.30.0.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
termcolor==3.1.0
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
